Move ThreadQueue tests to thread_operation folder
